Use explicitly typed integer literals (#43)
* Use explictly typed integer literals
* Java cannot cast int to Long (even though it is fine with int to long)
* Users might want to use a literal Long
* adds newline at end of file
* Add CI check for transpilation output comparison for examples

+++ examples/java/src/b/Loops.java
@@ -6,21 +6,21 @@ import java.util.ArrayList;
 public class Loops {
 public static final void main(String[] args) {
 long i = 0;
-while ((i < 10)) {
+while ((i < 10L)) {
 System.out.println(i);
 i = (i + 1);
 }
-ArrayList<Integer> tmp1 = new ArrayList<Integer>();
-tmp1.add(1);
-tmp1.add(2);
-tmp1.add(3);
-for (int ii : tmp1) {
+ArrayList<Long> tmp1 = new ArrayList<Long>();
+tmp1.add(1L);
+tmp1.add(2L);
+tmp1.add(3L);
+for (long ii : tmp1) {
 System.out.println(ii);
 }
 {
-int x = 0;
-while ((x < 10)) {
-x = (x + 1);
+long x = 0L;
+while ((x < 10L)) {
+x = (x + 1L);
 System.out.println(x);
 }
 }
